WIP: add some API to manage plugins and dependency paths
* change server port property type to ushort

+++ src/SMod3.Core/API/Objects/Server.cs
@@ -38,7 +38,7 @@ namespace SMod3.API
         /// <summary>
         ///     Server port.
         /// </summary>
-        public abstract int Port { get; }
+        public abstract ushort Port { get; }
 
         /// <summary>
         ///     Server ip address.
